panic

+++ k/execve/crts/libc/string.c
@@ -44,23 +44,33 @@ int strncmp(char *s1, char *s2, int len)
 {
 	int n = len;
 	
-	while(n > 0)
+	while( n > 0 )
 	{	
 	    n--;
         
-		if(*s1 != *s2){
-            return 1;
+		if( *s1 != *s2 )
+		{
+            return (int) 1;
         };
 		
 		*s1++;
 		*s2++;
+		
+		//#bugbug:
+		//Testar sem o asterísco, tanto em user mode(libc) quanto em kernel mode.
+		//s1++;
+		//s2++;
+		
+		
 	};		
 			
 			
-	if (*s1 != '\0' || *s2 != '\0' ){	
+	if (*s1 != '\0' || *s2 != '\0' )
+	{	
 	    return 2;
 	};
-	return 0;
+	
+	return (int) 0;
 };
